\name{pendenForm}
\alias{pendenForm}
\title{Formula interpretation and data transfer}
\description{
Function 'pendenForm' interprets the input 'form' of the function pencopula(),transfers the data back to the main program.
}
\usage{
pendenForm(penden.env)

}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{penden.env}{environment used in pendensity()}
}

\value{
Returning the values 'Y', the number of values 'n' and covariates 'p'.
}
\author{Christian Schellhase <cschellhase@wiwi.uni-bielefeld.de>}
\keyword{IO}
\references{Estimating Non-Simplified Vine Copulas Using Penalized Splines, Schellhase, C. and Spanhel, F. (2017), Statistics and Computing.}
