\name{print.pencopula}
\Rdversion{1.1}
\alias{print.pencopula}
\alias{print}

\title{Printing the main results of the penalized copula density estimation}

\description{Printing the call of the estimation, the used basis, lambda and
  the corresponding values of AIC and BIC. Need an object of class pencopula.}

\usage{
\method{print}{pencopula}(x, ...)
}

\arguments{
  \item{x}{x has to be object of class pencopula}
  \item{...}{further arguments}
}

\author{Christian Schellhase <cschellhase@wiwi.uni-bielefeld.de>}
\keyword{print}
\references{Estimating Non-Simplified Vine Copulas Using Penalized Splines, Schellhase, C. and Spanhel, F. (2017), Statistics and Computing.}
